Devolucion actualizacion busqueda wraper y nuevos metodos

+++ quski-oro-service-ejb/src/main/java/com/relative/quski/service/QuskiOroService.java
@@ -4428,7 +4428,7 @@ public class QuskiOroService {
 	}
 	private TbQoProceso createProcesoNegociacion( Long idReferencia, String usuario) throws RelativeException {
 		try {
-			TbQoProceso send = new TbQoProceso( BigDecimal.valueOf( idReferencia ));
+			TbQoProceso send = new TbQoProceso( Long.valueOf( idReferencia ));
 			send.setProceso( ProcesoEnum.NUEVO );
 			send.setEstadoProceso( EstadoProcesoEnum.CREADO );
 			send.setUsuario( usuario );
@@ -4440,7 +4440,7 @@ public class QuskiOroService {
 	}
 	private TbQoProceso createProcesoPago( Long idReferencia, String usuario) throws RelativeException {
 		try {
-			TbQoProceso send = new TbQoProceso( BigDecimal.valueOf( idReferencia ));
+			TbQoProceso send = new TbQoProceso( Long.valueOf( idReferencia ));
 			send.setProceso( ProcesoEnum.PAGO );
 			send.setEstadoProceso( EstadoProcesoEnum.PENDIENTE_APROBACION );
 			send.setUsuario( usuario );
